src/neural_network/requirements.txt
easyocr==1.7.1
Flask==3.0.3
imutils==0.5.4
keras_ocr==0.9.3
matplotlib==3.8.0
numpy==1.25.2
opencv_python==4.8.1.78
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pytesseract==0.3.10
torch==2.2.2
ultralytics==8.1.44

